Handle situation when searchQueryRules returns null (#591)

+++ Helper/Entity/ProductHelper.php
@@ -1037,6 +1037,10 @@ class ProductHelper
                 'hitsPerPage' => $hitsPerPage,
             ]);
 
+            if (!$fetchedQueryRules || !array_key_exists('hits', $fetchedQueryRules)) {
+                break;
+            }
+
             foreach ($fetchedQueryRules['hits'] as $hit) {
                 $index->deleteRule($hit['objectID'], true);
             }
